Add message.js
