+++ src/modules/interface.js
@@ -1,5 +1,6 @@
 import Phaser from 'phaser';
 import MenuItem from './menus/menuItem';
+import Message from './menus/message';
 
 // base menu class, container for menu items
 const Menu = new Phaser.Class({
@@ -79,54 +80,12 @@ const Menu = new Phaser.Class({
   },
 });
 
-// the message class extends containter
-export const Message = new Phaser.Class({
-
-  Extends: Phaser.GameObjects.Container,
-
-  initialize:
-  function Message(scene, events) {
-    Phaser.GameObjects.Container.call(this, scene, 160, 30);
-    const graphics = this.scene.add.graphics();
-    this.add(graphics);
-    graphics.lineStyle(1, 0xffffff, 0.8);
-    graphics.fillStyle(0x031f4c, 0.3);
-    graphics.strokeRect(-90, -15, 180, 30);
-    graphics.fillRect(-90, -15, 180, 30);
-
-    this.text = new Phaser.GameObjects.Text(scene, 0, 0, '', {
-      color: '#ffffff',
-      align: 'center',
-      fontSize: 13,
-      wordWrap: { width: 170, useAdvancedWrap: true },
-    });
-    this.add(this.text);
-    this.text.setOrigin(0.5);
-    events.on('Message', this.showMessage, this);
-    this.visible = false;
-  },
-  showMessage(text) {
-    this.text.setText(text);
-    this.visible = true;
-    if (this.hideEvent) {
-      this.hideEvent.remove(false);
-    }
-    this.hideEvent = this.scene.time.addEvent(
-      { delay: 2000, callback: this.hideMessage, callbackScope: this },
-    );
-  },
-  hideMessage() {
-    this.hideEvent = null;
-    this.visible = false;
-  },
-});
-
 const HeroesMenu = new Phaser.Class({
   Extends: Menu,
 
   initialize:
 
-  function HeroesMenu(x, y, scene) {
+  function HeroesMenu(x,s y, scene) {
     Menu.call(this, x, y, scene);
   },
 });
